Refactor tests to share one single list of tests and extend from basic test class
